Fixed the background color of the tabs + made PiShockService into a static service + Properly made the Shockers tab a tab + Added the icon near the title + Various fixes
