+++ TurretShockyUI/TurretShocky.csproj
@@ -6,8 +6,8 @@
     <BuiltInComInteropSupport>true</BuiltInComInteropSupport>
     <ApplicationManifest>app.manifest</ApplicationManifest>
     <AvaloniaUseCompiledBindingsByDefault>true</AvaloniaUseCompiledBindingsByDefault>
-    <AssemblyVersion>1.3.0.5211</AssemblyVersion>
-    <FileVersion>1.3.0.5211</FileVersion>
+    <AssemblyVersion>1.3.0.5220</AssemblyVersion>
+    <FileVersion>1.3.0.5220</FileVersion>
     <ApplicationIcon>Assets\Shocky.ico</ApplicationIcon>
     <AssemblyName>TurretShocky</AssemblyName>
   </PropertyGroup>
